(hana) emme2 path fixed
git-svn-id: https://svn.urbansim.org/trunk@12872 1513fa5d-83ce-422e-ae47-6967bb0fb594

+++ opus_emme2/models/restore_trip_tables.py
@@ -22,8 +22,9 @@ class RestoreTripTables:
     def run(self, config, source_directory, year):
         dir = config['travel_model_configuration'][year]['bank'][0]
         dst = os.path.join(os.environ['TRAVELMODELROOT'], dir, 'triptabs')
-        src = os.path.join(dir, source_directory)
-        shutil.rmtree(dst)
+        src = os.path.join(os.environ['TRAVELMODELROOT'], dir, source_directory)
+        if os.path.exists(dst):
+            shutil.rmtree(dst)
         shutil.copytree(src, dst)
         
 if __name__ == "__main__":
